Add files via upload
This is a cluster of multi-magnetic center. In addition, there are two different Mn ions with difference, either Mn+3 (S=3/2) or Mn+4 (S=2). It will be important that LAPW can determined it automatically.
  
ps. The Mn_taa molecule should have S=1 as ground state and S=2 as the first excited state.